Restructure map commands
